+++ syncfit/src/App.jsx
@@ -5,7 +5,6 @@ import Chat from './pages/Chat';
 import Contact from './pages/Contact';
 import Login from './pages/Login'
 import Notfound from './pages/NotFound'
-import NavBar from './components/MobileNav'
 
 const router = createBrowserRouter([
   {
