+++ syncfit/src/components/Layout/Footer/Footer.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import '../../css/Footer.css'
+
+const Footer = () => {
+  return (
+    <div className='footer'>
+      <h4>© SyncFit Music, 2025. All rights reserved.</h4>
+    </div>
+  )
+}
+
+export default Footer
