+++ syncfit/src/components/Layout/Header/Nav.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { Link, useLocation } from 'react-router-dom';
+import '../../css/Nav.css';
+
+const links = [
+  { name: 'Home', path: '/' },
+  { name: 'Chat', path: '/chat' },
+  { name: 'Contact', path: '/contact' },
+  { name: 'Login', path: '/login' },
+];
+
+const Nav = () => {
+  const location = useLocation(); 
+
+  return (
+    <nav className="nav">
+      {links.map((link, index) => (
+        <Link
+          to={link.path}
+          key={index}
+          className={`${link.path === location.pathname ? 'nav-link-active' : 'nav-link'}`}
+        >
+          {link.name}
+        </Link>
+      ))}
+    </nav>
+  );
+};
+
+export default Nav;
