+++ syncfit/src/components/css/Footer.css
@@ -1,0 +1,3 @@
+.footer {
+  text-align: center;
+}
