+++ syncfit/src/components/css/MobileNav.css
@@ -11,7 +11,7 @@
 
 .NavIcon {
     font-size: 28px;
-    color: black;
+    color: white;
 }
   
 .SideBar {
