+++ syncfit/src/pages/Chat.jsx
@@ -1,8 +1,11 @@
 import React from 'react'
+import Layout from '../components/Layout/Layout'
 
 const Chat = () => {
   return (
-    <div>Chat</div>
+    <Layout>
+      Chat
+    </Layout>
   )
 }
