+++ syncfit/src/pages/Contact.jsx
@@ -1,8 +1,11 @@
 import React from 'react'
+import Layout from '../components/Layout/Layout'
 
 const Contact = () => {
   return (
-    <div>Contact</div>
+    <Layout>
+      Contact
+    </Layout>
   )
 }
