+++ syncfit/src/pages/Login.jsx
@@ -2,7 +2,9 @@ import React from 'react'
 
 const Login = () => {
   return (
-    <div>Login</div>
+    <Layout>
+      Login
+    </Layout>
   )
 }
