+++ syncfit/src/pages/css/NotFound.css
@@ -1,0 +1,6 @@
+.notfound {
+  width: 100vw;
+  height: 100vh;
+  background-color: #FFD100;
+  
+}
